extract SignedMessagesHelpLink to a different file
